requirements.txt
# Core utilities
numpy>=1.25.0
pandas>=2.0.0

# Streamlit app
streamlit>=1.28.0

# Resume parsing
pypdf>=3.12.0
python-docx>=1.0.0

# LLM clients
openai>=1.6.0
google-generativeai>=0.3.2

# Web scraping (Mercor, Selenium)
selenium>=4.15.0

# Machine learning & clustering
scikit-learn>=1.3.0

# Visualization
matplotlib>=3.8.0
seaborn>=0.12.2